+++ SVD_matrix.h
@@ -9,7 +9,7 @@
 void SVD_matrix_mul(matrix matrix1, matrix matrix2, matrix result);
 void SVD_matrix_trans(matrix in, matrix out);
 void SVD_matrix_copy(matrix in, matrix out);
-float SVD_matrix_dot(matrix matrix1, matrix matrix2, int row1, int col2);
+int32_t SVD_matrix_dot(matrix matrix1, matrix matrix2, int row1, int col2);
 void SVD_matrix_rotation_angles(matrix matrix);
 bool SVD_matrix_equal(matrix matrix1, matrix matrix2);
 bool SVD_matrix_isDiagonal(matrix in);
